+++ CHANGELOG.md
@@ -5,7 +5,20 @@ The sections should follow the order `Packaging`, `Added`, `Changed`, `Fixed` an
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## 0.8.0-dev
+## 0.9.0-dev
+
+### Added
+
+- Support for `ipfs`/`ipns` URLs
+- Mode field for regex hint bindings
+
+### Fixed
+
+- Regression in rendering performance with dense grids since 0.6.0
+- Crash/Freezes with partially visible fullwidth characters due to alt screen resize
+- Incorrect vi cursor position after invoking `ScrollPageHalfUp` action
+
+## 0.8.0
 
 ### Packaging